pyproject.toml
[project]
name = "pvt-srp-delta-breakout-personal-trade"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=0.19.0",
    "requests>=2.25.0",
]
